+++ framework/console/controllers/CacheController.php
@@ -121,6 +121,43 @@ class CacheController extends Controller
     }
 
     /**
+     * Clears cache scheme for given component connection to database
+     * For example,
+     *
+     * ~~~
+     * # clears cache scheme specified by component id: "db"
+     * yii cache/clear-schema db
+     * ~~~
+     *
+     * @param string $db id connection component
+     * @throws Exception
+     * @throws \yii\base\InvalidConfigException
+     */
+    public function actionClearSchema($db = 'db')
+    {
+        $connection = $connection = Yii::$app->get($db, false);
+        if (is_null($connection)) {
+            $this->stdout("$db - unknown component\n", Console::FG_RED);
+            return self::EXIT_CODE_ERROR;
+        }
+
+        if (!$connection instanceof \yii\db\Connection) {
+            $this->stdout("$db - This component doesn't inherit \\yii\\db\\Connection\n", Console::FG_RED);
+            return self::EXIT_CODE_ERROR;
+        } else if (!$this->confirm("\nClear cache schema for '$db' connection?")) {
+            return static::EXIT_CODE_NORMAL;
+        }
+
+        try {
+            $schema = $connection->getSchema();
+            $schema->refresh();
+            $this->stdout("Schema cache for component $db, was cleared\n\n", Console::FG_GREEN);
+        } catch (\Exception $e) {
+            $this->stdout($e->getMessage() . "\n\n", Console::FG_RED);
+        }
+    }
+
+    /**
      * Notifies user that given caches are found and can be flushed.
      * @param array $caches array of cache component classes
      */